% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{apmc}
\alias{apmc}
\title{A Wraper of glht to Get All Pairwise Mean Comparisons}
\usage{
apmc(X, model, focus, test = "single-step", level = 0.05)
}
\arguments{
\item{X}{a matrix where each line is a linear function of the model
parameters to estimate a least squares mean. In most pratical
cases, it is an object from the \code{\link[doBy]{LSmatrix}}.}

\item{model}{a model with class recognized by
\code{\link[multcomp]{glht}}.}

\item{focus}{a string with the name of the factor which levels are
being compared.}

\item{test}{a p-value correction method. See
\code{\link[stats]{p.adjust.methods}}.}

\item{level}{the experimentwise significance level for the multiple
comparisons. The individual coverage of the confidence interval
is \code{1-level}. Default is 0.05.}
}
\value{
a \code{data.frame} with interval estimates and compact
    letter display for the means comparisons.
}
\description{
This function performs all pairwise compararisons among
    means returning pontual and intervalar estimates followed by
    letters to easy discriminate values. It is in fact a wraper of
    \code{\link[multcomp]{glht}}.
}
\examples{

library(doBy)
library(multcomp)

# Single factor.
m0 <- lm(weight ~ feed, data = chickwts)
anova(m0)

# Prepare the matrix to estimate lsmeans.
X <- LSmatrix(m0, effect = "feed")
rownames(X) <- levels(chickwts$feed)
apmc(X, model = m0, focus = "feed", test = "fdr")

# Two factors.
m1 <- lm(breaks ~ wool * tension, data = warpbreaks)
anova(m1)

X <- LSmatrix(m1, effect = c("wool", "tension"))
attributes(X)

Xs <- by(X, INDICES = attr(X, "grid")$tension, FUN = as.matrix)
Xs <- lapply(Xs, "rownames<-", levels(warpbreaks$wool))

# Comparing means of wool in each tension.
lapply(Xs, apmc, model = m1, focus = "wool",
       test = "single-step", level = 0.1)

}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}
\seealso{
\code{\link{apc}}, \code{\link[doBy]{LSmatrix}},
    \code{\link[multcomp]{glht}}.
}

