% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.lm.R
\name{as.lm}
\alias{as.lm}
\alias{as.lm.nls}
\title{Converts nls class objects to lm class}
\usage{
as.lm(object, ...)

\method{as.lm}{nls}(object, ...)
}
\arguments{
\item{object}{An object of class \code{nls}.}

\item{...}{Currently not used.}
}
\value{
This function returns an object of class \code{lm} created
    from the \code{nls} object. The linear model design matrix used
    is the the partial derivatives of the \code{nls} model function
    with relation to the model parameters, i.e. the gradient matrix.
}
\description{
This functions is a modified version of the code
    available at the \code{nls2} package on github
    \url{https://github.com/ggrothendieck/nls2}. This function is no
    longer available in the current version of \code{nls2} package.
}
\details{
This function is useful to get the residuals plot for the
    fitted nls model. The dependence on the \code{as.proto.list}
    function in the \code{proto} package was removed (also this
    function no longer exists). Thanks for the original author of
    this function G. Grothendieck.
}
\examples{

# An simple nls fit.
n0 <- nls(dist ~ A + B * speed^C,
          data = cars,
          start = c(A = 0, B = 1, C = 1))
summary(n0)

# The results.
plot(dist ~ speed, data = cars)
with(as.list(coef(n0)), {
    curve(A + B * speed^C, xname = "speed", add = TRUE)
})

# Residual analysis.
par(mfrow = c(2, 2))
plot(as.lm(n0))
layout(1)

}
\author{
Walmes Zeviani, \email{walmes@ufr.br}
}

