% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Area under curve by trapezoidal integration}
\usage{
auc(time, resp)
}
\arguments{
\item{time}{A vector with the time numeric variable.}

\item{resp}{A vector with the response numeric variable.}
}
\value{
Returns the area under the curve based on the trapezoidal
    method.
}
\description{
Calculates the area under curve based on the trapezoidal
    integration method.
}
\examples{

x <- 1:5
y <- c(1, 2, 3, 2, 1)

plot(y ~ x,
     type = "o",
     ylim = c(0, max(y)),
     asp = 1)
polygon(x = c(x, rev(x)),
        y = c(y, 0 * y),
        border = NA,
        col = "purple")
abline(v = x, h = x, lty = 2, col = "gray")

# Area for the colored region.
auc(time = x, resp = y)

}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}
