% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.groups.segplot.R
\name{centfac}
\alias{centfac}
\title{Numeric Centered Factor Levels}
\usage{
centfac(group, space = NULL)
}
\arguments{
\item{group}{A factor.}

\item{space}{A numeric value to be used as the space between
levels. If \code{NULL}, the space is determined by the
\code{group}.}
}
\description{
This function receveis a factor and return a numeric
    vector with equally spaced factor levels centered at 0.
}
\examples{

centfac(warpbreaks$tension)
centfac(warpbreaks$tension, space = 1)
centfac(warpbreaks$wool)
centfac(warpbreaks$wool, space = 1)

}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}

