% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{cld2}
\alias{cld2}
\title{Modified Compact Letter Display to Irregular Designs}
\usage{
cld2(object, level = 0.05)
}
\arguments{
\item{object}{an object returned by \code{\link[multcomp]{glht}}. It
is assumed that the matrix used as the \code{linfct} argument in
\code{glht} corresponds to a matrix to get Tukey contrasts of
least squares means.}

\item{level}{the nominal significance level.}
}
\value{
an object of class \code{"cld"} with letters to resume mean
    comparisons.
}
\description{
This functions get the compact letter display for
    objects of class \code{"glht"}. Modification was done to get the
    letters to design with missing cells, non completelly crossed
    factorial designs and nested factorial designs. These models are
    usually declared by a model matrix to have all effects
    estimated. It is assumed that Tukey contrasts was used.
}
\examples{

# Toy data 1: experiment with cultivars in several locations.
td1 <- expand.grid(loc = gl(5, 1),
                   block = gl(3, 1),
                   cult = LETTERS[1:6])
td1 <- subset(td1, !(loc == 1 & cult == "A"))
td1 <- subset(td1, !(loc == 2 & cult == "B"))
xtabs(~loc + cult, td1)
td1$y <- seq_len(nrow(td1))

library(lme4)

# Fit the mixed model.
m0 <- lmer(y ~ loc * cult + (1 | loc:block), data = td1)
logLik(m0)

# The same model but without rank deficience.
td1$loccult <- with(td1, interaction(loc, cult, drop = TRUE))
m1 <- lmer(y ~ loccult + (1 | loc:block), data = td1)
logLik(m1)

library(doBy)

X <- LSmatrix(lm(nobars(formula(m1)), data = td1), effect = "loccult")
rownames(X) <- levels(td1$loccult)
dim(X)

Xs <- X[grepl(x = rownames(X), "^1\\\\."),]
Xc <- apc(Xs)

library(multcomp)

g <- summary(glht(m1, linfct = Xc), test = adjusted(type = "fdr"))

cld2(g)

confint(glht(m1, linfct = Xs), calpha = univariate_calpha())

}
\author{
Walmes Zeviani, \email{walmes@ufpr.r}.
}
\seealso{
\code{\link{apc}}, \code{\link[doBy]{LSmatrix}},
    \code{\link[multcomp]{glht}}.
}

