% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_tree.R
\name{dir_tree}
\alias{dir_tree}
\title{Shows directories content as an ASCII tree graph}
\usage{
dir_tree(path = "./", level = 3, type = "dir_last", all.files = FALSE)
}
\arguments{
\item{path}{character[1] Root path for the construction of the ASCII
representation.}

\item{level}{integer[1] A positive integer that is how many levels to
show. Default is 3. Use \code{Inf} to not limit.}

\item{type}{character[1] This is the display mode. Default is
\code{"dir_last"} that keeps files at the beginning and
directories to the end. The value \code{"none"} uses alphabetical
order..}

\item{all.files}{logical[1] Using \code{TRUE}, hidden files and
directories are showed.}
}
\value{
Returns the directory list as invisible.
}
\description{
This function mimics the
    \href{http://mama.indstate.edu/users/ice/tree/}{\code{tree}}
    shell command in Linux. It shows directories in a tree structure
    as an ASCII art. Not all options were enabled.
}
\examples{

\dontrun{

# Tree of your currend directory.
getwd()
dir_tree()
dir_tree(type = "none")
dir_tree(all.files = TRUE)

# Tree of a installed package.
dir_tree(system.file(package = "lattice"), level = 1)
dir_tree(system.file(package = "lattice"), level = 2)
dir_tree(system.file(package = "lattice"), level = 3)
dir_tree(system.file(package = "lattice"), level = 4)

m <- dir_tree(system.file(package = "MASS"), level = 2)
str(m)
}
}
\seealso{
\code{\link{tree}()} for an implementation that calls
    \code{tree} on Linux. Also, \code{\link[base]{dir}()},
    \code{\link[base]{list.files}()} and
    \code{\link[base]{list.dirs}()} can be useful in many
    circumstances.
}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}
