% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enc_titles.R
\name{enc_titles}
\alias{enc_titles}
\alias{use_enc}
\alias{enc_nize}
\title{Applies \code{enc{}{}} to non ASCII words}
\usage{
use_enc(x)

enc_nize(file)

enc_titles()
}
\arguments{
\item{x}{character[1] A title to be converted.}

\item{file}{character[1] A R file with documentation written in
roxygen syntax. \code{enc_nize()} assumes that title is signed
with \code{@title} tag. If it is not the case, the function give
a warning. This function returns \code{NULL} and modifies files
in disc. Be careful with its use.}
}
\value{
character[1] The converted character provided.
}
\description{
Non ASCII words (commom in portuguese) trigers warnings
    in \code{R CMD CHECK}. To prevent this, \code{enc{}{}} must be
    used. This function gets a character and returns \code{enc{}{}}
    for all non ASCII words.
}
\details{
\code{enc_titles} applies \code{enc_nize} to all files in
    the \code{./R/} directory of a package. So, it is assumed that
    working directory is a root package directory. \code{enc_titles}
    modifies files. So be careful with its use.
}
\examples{

library(proto)
library(gsubfn)

cat(use_enc("Um título com conteúdo acenturado para testar a função"), "\n")

}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}
