% Generated by roxygen2 (4.0.2): do not edit by hand
\name{equallevels}
\alias{equallevels}
\title{Make that to data frames have the same factor levels}
\usage{
equallevels(x, y)
}
\arguments{
\item{x}{the target data frame.}

\item{y}{the reference data frame.}
}
\value{
the restructred data frame.
}
\description{
If two data frames have factor ou character columns with
the same name, those in the first will have the same level order as
those in the second. So, in terms of factor columns, these data
frames will have the same levels in the same order. Character columns
in the first will be converted to factor.
}
\examples{
\donttest{
a <- data.frame(Species=as.character(
                    sample(iris$Species, 10, repl=TRUE)),
                stringsAsFactors=FALSE)
str(a)
b <- equallevels(a, iris)
str(b)
}
}

