% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{matrix2html}
\alias{matrix2html}
\title{Convert a character matrix to HTML table}
\usage{
matrix2html(x, caption = NULL, styles = NULL, indexes = NULL,
  class = NULL)
}
\arguments{
\item{x}{a character matrix. Is most of the cases it is the object
returned by \code{table_format}.}

\item{caption}{a string to be the caption of the table. Default is
\code{NULL}, no caption.}

\item{styles}{a vector with style definition for \code{th}, \code{tr}
and \code{td} tags. See the examples section. Default is
\code{NULL}. These styles must be defined in a css file used to
render the html page. If you are using \code{knitr}, you can use
a custom css file when knitting \code{knit2html("my\_doc.Rmd",
stylesheet = "my\_custom\_css.css")}.}

\item{indexes}{a positive integer matrix with dimensions equal to
\code{x}. The numbers in each cell of \code{indexes} call the
corresponding \code{styles} to be used in \code{x}. See examples
for clarification. Default is \code{NULL}.}

\item{class}{a string corresponding a table style defined in a
\code{css} file. Default is \code{NULL}.}
}
\value{
a character vector. Use \code{cat} inside chunks with header
    \code{results = "asis"} to print the result as interpretable code
    of an html table.
}
\description{
This function coverts a character matrix to an html
    table.
}
\examples{

x <- head(rock)
x <- sapply(x, as.character)
str(x)

m2h <- matrix2html(x)
cat(m2h)

x <- mtcars[1:10, 1:3]
x <- rbind(c("", colnames(x)), cbind(rownames(x), x))
x <- as.matrix(x)
rownames(x) <- colnames(x) <- NULL
x[1,1] <- "Cars"
x

m2h <- matrix2html(x, caption="Part of the cars data set.")
cat(m2h)

# These table class are assumed to be defined in a css file as
# follow.

# td.red { color: #CC0000; }
# td.blue { color: #0000CC; }
# td.shade { background-color: #CCCCCC; }
# td.line { border-bottom: 1px solid #000000; }
# td.bold { font-weight: bold; }
# td.italic { font-style: italic; }

sty <- c("<th align=\\"center\\">\%s</th>", "<tr>\\n\%s\\n</tr>\\n",
         "<td align=\\"center\\">\%s</td>",
         "<td align=\\"center\\" class=\\"red\\">\%s</td>",
         "<td align=\\"center\\" class=\\"blue line\\">\%s</td>",
         "<td align=\\"center\\" class=\\"shade\\">\%s</td>",
         "<td align=\\"center\\" class=\\"line\\">\%s</td>",
         "<td align=\\"center\\" class=\\"bold shade\\">\%s</td>",
         "<td align=\\"center\\" class=\\"italic blue\\">\%s</td>")

# Which style for which cell table.
idx <- 0L * row(x) + 3L
idx[3, ] <- 4L
idx[4, ] <- 5L
idx[5, ] <- 6L
idx[7, ] <- 7L
idx[8, ] <- 8L
idx[9, ] <- 9L
idx[, 3] <- 3L
idx[1, ] <- 1L

m2h <- matrix2html(x = x, styles = sty, indexes = idx,
                   caption = "Part of the cars data set.")
cat(m2h)

}
\seealso{
\code{\link{table_format}()}.
}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}
