% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_in_gnumeric.R
\name{open_in_gnumeric}
\alias{open_in_gnumeric}
\title{Open data frames with Gnumeric}
\usage{
open_in_gnumeric(tb)
}
\arguments{
\item{tb}{A data frame.}
}
\value{
None. This function opens the data frame with Gnumeric
}
\description{
Takes a data frame, saves in a temporary file and opens
    it with Gnumeric spread sheet software. This function only works
    on Linux and requires Gnumeric installed. It can be modified to
    be used with another software and operating system. The code is
    based on this answer \url{https://stackoverflow.com/a/28580683}
    on StackOverflow.
}
\examples{

\dontrun{

open_in_gnumeric(iris)
open_in_gnumeric(PlantGrowth)
open_in_gnumeric(swiss)

}

}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}
