% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{ordered_cld}
\alias{ordered_cld}
\title{Order Letters According to Numeric Vector}
\usage{
ordered_cld(let, means = let)
}
\arguments{
\item{let}{Character vector with the letters returned by
\code{\link[multcomp]{cld}()} or \code{\link{cld2}()}.}

\item{means}{Numeric vector with the corresponding estimates in which
the highest value will have the letter \code{a}.}
}
\value{
A character vector with the letters rearranged.
}
\description{
This function order the letters in the compact letter
    display to the highest estimate receive the letter \code{a}. This
    is a convetion in most software for analysis of experiments.
}
\examples{

# Toy data.
set.seed(4321)
td <- data.frame(trt = rep(sample(1:8), each = 5))
td$y <- rnorm(nrow(td), mean = sort(td$trt), sd = 2)

plot(y ~ trt, data = td)

# Fit the model.
td$trt <- factor(td$trt)
m0 <- lm(y ~ trt, data = td)
anova(m0)
summary(m0)

library(multcomp)
library(doBy)

X <- LE_matrix(m0, effect = "trt")
rownames(X) <- levels(td$trt)
Xc <- apc(X)

g <- summary(glht(m0, linfct = Xc),
             test = adjusted(type = "fdr"))

res <- data.frame(trt = levels(td$trt),
                  mean = X \%*\% coef(m0))

let <- cld2(g)
res$cld2 <- let$mcletters$Letters
res[order(res$mean, decreasing = TRUE), ]

res$let2 <- ordered_cld(res$cld2, res$mean)
res[order(res$mean, decreasing = TRUE), ]

\dontrun{

library(latticeExtra)
library(grid)

ci <- as.data.frame(
    confint(glht(m0, linfct = X),
            calpha = univariate_calpha())$confint)
ci <- cbind(res, ci)

segplot(reorder(trt, Estimate) ~ lwr + upr,
        centers = Estimate,
        data = ci,
        draw = FALSE,
        cld = ci$let2,
        par.settings = list(layout.widths = list(right.padding = 7))) +
    layer(panel.text(x = centers,
                     y = z,
                     labels = sprintf("\%0.2f \%s",
                                      centers,
                                      cld),
                     pos = 3))

ocld <- with(ci[order(ci$Estimate), ],
     ordered_cld(cld2, Estimate))
x <- attr(ocld, "ind")
index <- which(x, arr.ind = TRUE)
trellis.focus("panel", column = 1, row = 1, clip.off = TRUE)
xcor <- 1.03 + (index[, 2] - 1)/50
grid.segments(x0 = unit(xcor, "npc"),
              x1 = unit(xcor, "npc"),
              y0 = unit(index[, 1] + 0.5, units = "native"),
              y1 = unit(index[, 1] - 0.5, units = "native"),
              gp = gpar(lwd = 2, col = "blue"))
trellis.unfocus()

}

}
\seealso{
\code{\link{cld2}()}.
}
\author{
Walmes Zeviani, \email{walmes@ufpr.r}.
}
