% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{ordered_cld2}
\alias{ordered_cld2}
\alias{ordered_cld}
\title{Order Letters According to Numeric Vector}
\usage{
ordered_cld(let, means = let)
}
\arguments{
\item{let}{Character vector with the letters returned by \code{cld2}.}

\item{means}{Numeric vector with the corresponding estimates in which
the highest value will have the letter \code{a}.}
}
\value{
A character vector with the letters rearranged.
}
\description{
This function order the letters in the compact letter
    display to the highest estimate receive the letter \code{a}. This
    is a convetion in most software for analysis of experiments.
}
\examples{

# Toy data.
set.seed(4321)
td <- data.frame(trt = rep(sample(1:8), each = 5))
td$y <- rnorm(nrow(td), mean = sort(td$trt), sd = 2)

plot(y ~ trt, data = td)

# Fit the model.
td$trt <- factor(td$trt)
m0 <- lm(y ~ trt, data = td)
anova(m0)
summary(m0)

library(multcomp)
library(doBy)

X <- LSmatrix(m0, effect = "trt")
rownames(X) <- levels(td$trt)
Xc <- apc(X)

g <- summary(glht(m0, linfct = Xc),
             test = adjusted(type = "fdr"))

res <- data.frame(trt = levels(td$trt),
                  mean = X \%*\% coef(m0))

let <- cld2(g)
res$cld2 <- let$mcletters$Letters
res[order(res$mean), ]

res$let2 <- ordered_cld(res$cld2, res$mean)
res[order(res$mean), ]

}
\seealso{
\code{\link{cld2}()}.
}
\author{
Walmes Zeviani, \email{walmes@ufpr.r}.
}
