% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/panel.beeswarm.R
\name{panel.beeswarm}
\alias{panel.beeswarm}
\title{Lattice panel for scatter plots with a beeswarm flavor}
\usage{
panel.beeswarm(x, y, subscripts, spread, ...)
}
\arguments{
\item{x,y,subscripts}{arguments passed to panel.xyplot, the core
function of xyplot.}

\item{spread}{a numeric scalar indicating how much spread the
values. Should be determined by try and error. If you change the
window plot dimensions, you should determine a new value. As an
advice, always use a \code{x11()} before the plot you want to
guarantee a good look.}

\item{...}{arguments passed to panel.xyplot, the core function of
xyplot.}
}
\value{
None is returned.
}
\description{
Used to make scatter plot of discrete variables with no
overlapping points. Observations with the same y value are spread.
}
\examples{
\donttest{
##-----------------------------------------------------------------------------
## A simple example.

est.ing <- c("vegetative","flower bud","blossom","fig","cotton boll")
cap <- expand.grid(rept=1:5,
                   des=seq(0,100,l=5)/100,
                   est=factor(est.ing, levels=est.ing))

cap$nc <- c(10, 9, 8, 8, 10, 11, 9, 10, 10, 10, 8, 8, 10, 8, 9, 9, 7, 7,
            8, 9, 8, 6, 6, 5, 6, 7, 8, 8, 9, 10, 9, 12, 7, 10, 9, 8, 9,
            9, 10, 8, 11, 10, 7, 8, 8, 7, 7, 7, 7, 8, 10, 9, 8, 12, 8,
            7, 5, 5, 7, 5, 6, 5, 7, 4, 7, 8, 5, 7, 6, 4, 5, 5, 4, 4, 5,
            8, 10, 7, 8, 10, 9, 6, 6, 8, 6, 9, 7, 11, 8, 9,6, 6, 6, 6,
            7, 3, 3, 2, 4, 3, 11, 7, 9, 12 , 11, 9, 13, 8, 10, 10, 9, 7,
            7, 9, 9, 8, 8, 10, 8, 10, 9, 8, 10, 8, 10)

require(lattice)

x11(width=7, height=2.8)
xyplot(nc~des|est, data=cap,
       layout=c(5,1), type=c("p","smooth"),
       as.table=TRUE, col=1, col.line="gray50",
       xlab="Artificial defoliation level",
       ylab="Number of bolls produced",
       strip=strip.custom(bg="gray90", factor.levels=est.ing),
       xlim=extendrange(c(0:1), f=0.15),
       spread=0.07, panel=panel.beeswarm)

x11(width=7, height=2.8)
xyplot(nc~est|factor(des), data=cap,
       layout=c(5,1), type=c("p","smooth"),
       as.table=TRUE, col=1, col.line="gray50",
       xlab="Cotton growth stage",
       ylab="Number of bolls produced",
       strip=strip.custom(bg="gray90"),
       xlim=extendrange(c(1, nlevels(cap$est)), f=0.2),
       scales=list(
           x=list(at=1:nlevels(cap$est),
           labels=substr(levels(cap$est), 1, 3))),
       spread=0.35, panel=panel.beeswarm)

}
}
\seealso{
\code{\link[lattice]{xyplot}}.
}

