% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.groups.segplot.R
\name{panel.groups.segplot}
\alias{panel.groups.segplot}
\title{Panel to Plot Confidence Intervals by Groups in
    \code{segplot()}}
\usage{
panel.groups.segplot(x, y, z, centers, groups, gap = NULL, data,
  subscripts, ...)
}
\arguments{
\item{x, y, z, centers, data, subscripts, ...}{Arguments passed to
\code{\link[latticeExtra]{segplot}()}.}

\item{groups}{The grouping variable (\code{factor}), with the same
length of \code{lwr} e \code{upr}.}

\item{gap}{Scalar that is the distance among segments. Default is
0.1. If the grouping factor has \eqn{k} levels, so \eqn{0 \leq
\textrm{gap} < 1/k}. A negative value for \code{gap} will put the
segments in a reversed order.}
}
\description{
This function allows non overlapping error bars in
    \code{latticeExtra::segplot()}. It has the \code{groups =}
    argument.
}
\examples{

library(latticeExtra)

m0 <- lm(log(breaks) ~ wool * tension, data = warpbreaks)

pred <- with(warpbreaks, expand.grid(KEEP.OUT.ATTRS = TRUE,
                                     wool = levels(wool),
                                     tension = levels(tension)))

pred <- cbind(pred,
              predict(m0, newdata = pred, interval = "confidence"))
str(pred)

segplot(wool ~ lwr + upr, centers = fit, data = pred,
        draw = FALSE, horizontal = FALSE)

segplot(wool ~ lwr + upr, centers = fit, data = pred,
        draw = FALSE, horizontal = FALSE,
        groups = tension, gap = 0.05,
        panel = panel.groups.segplot)

}
\seealso{
\code{\link[latticeExtra]{segplot}()}.
}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}
