% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/panel.segplotBy.R
\name{panel.segplotBy}
\alias{panel.segplotBy}
\title{Lattice panel to non overlapping segments in \code{segplot()}}
\usage{
panel.segplotBy(x, y, z, centers, subscripts, groups, f, ...)
}
\arguments{
\item{x,y,z,centers,subscripts,...}{see \code{\link[latticeExtra]{segplot}}.}

\item{groups}{the grouping variable. Must be a factor.}

\item{f}{factor that is the vertical distance among arrows.}
}
\value{
None is returned.
}
\description{
This panel allows no overlapping of segments in
\code{latticeExtra::segplot()}.
}
\examples{
require(doBy)
require(multcomp)
require(plyr)
require(latticeExtra)

m0 <- lm(breaks~wool*tension, data=warpbreaks)
X <- LSmatrix(m0, effect=c("wool", "tension"))
attributes(X)

Xs <- by(X, INDICES=attr(X, "grid")$tension, FUN=as.matrix)
Xs <- lapply(Xs, "rownames<-", levels(warpbreaks$wool))
L <- lapply(Xs, apmc, model=m0, focus="wool", test="single-step",
            level=0.10)

L <- ldply(L)
names(L)[1] <- "tension"
L$tension <- factor(L$tension, levels=levels(warpbreaks$tension))
str(L)

segplot(wool~lwr+upr, centers=estimate, data=L, draw=FALSE)

segplot(tension~lwr+upr, centers=estimate, data=L, draw=FALSE,
        panel=panel.segplotBy, groups=L$wool, f=0.05)
}
\seealso{
\code{\link[latticeExtra]{segplot}}
}

