% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs_versions.R
\name{pkgs_versions}
\alias{pkgs_versions}
\title{Returns the fields with packages in DESCRIPTION with versions}
\usage{
pkgs_versions(pkg)
}
\arguments{
\item{pkg}{\code{character[1]} Name of an available package.}
}
\value{
When assigned, the return value is a character that can be
    properly display with \code{cat()}.
}
\description{
This function returns the packages mentioned in the
    DESCRIPTION file of a package and returns the current versions of
    them appended after the package name. So, you can copy and paste
    directly in the DESCRIPTION file to inform the minimal versions
    required for each dependence. The versions appended are those of
    the package in the current environment. So, be aware of using the
    correct packages version to develop packages and inform version
    only of those packages that are critical for the functionality of
    yours.
}
\examples{

\dontrun{

pkgs_versions("lattice")

devtools::load_all("~/Projects/wzRfun/")
pkgs_versions("wzRfun")

devtools::load_all("~/Projects/labestData")
u <- pkgs_versions("labestData")
cat(u)

}
}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}
