% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2nls.R
\name{R2nls}
\alias{R2nls}
\title{Coeficient of Determination for Non Linear Regression Models}
\usage{
R2nls(nls.obj)
}
\arguments{
\item{nls.obj}{a object of class \code{nls}.}
}
\value{
a list with two slots. The first named \code{anova} is the
    table with partition of the total sum of squares (adjusted for
    the mean or null model or intercept inly model) in regression sum
    of squares and residual sum of squares. The second named
    \code{R2} is the coefficient of determination.
}
\description{
This function returns the anova table with the partition
    of the total sum of squares into due the regression model and
    residuals and the corresponding \eqn{R^2} (adjusted for the null
    model).
}
\examples{

da <- subset(DNase, Run == 4)
plot(density ~ log(conc), data = da)

m0 <- nls(density ~ Asym/(1 + exp(-(log(conc) - xmid)/scal)),
          start = list(Asym = 2.3, xmid = 1.5, scal = 1),
          data = da)
summary(m0)

plot(density ~ log(conc), data = da)
with(as.list(coef(m0)),
     curve(Asym/(1 + exp(-(x - xmid)/scal)),
           add = TRUE, col = 2))

R2nls(m0)
cor(fitted(m0), da$density)^2

}
\author{
Walmes Zeviani, \email{walmes@ufr.br}
}

