% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{radial_cld}
\alias{radial_cld}
\title{Radial Plot for a Compact Letter Display Vector}
\usage{
radial_cld(
  cld,
  labels = cld,
  col = NULL,
  means = NULL,
  perim = FALSE,
  legend = TRUE
)
}
\arguments{
\item{cld}{Character vector with strings of letters that indicates
which pair of treatment cells are not different.}

\item{labels}{Vector of text to be annotated next each point.}

\item{col}{Vector of colors to be used in the segments that joint
points.}

\item{means}{Numeric vector with the estimated means of treatment
cells. It is used to place points at distances proportional to
the differences on means.}

\item{perim}{Logical value (default is \code{FALSE}) that indicates
weather draw or not a circle in the perimeter passing by the
points.}

\item{legend}{Logical value (default is \code{TRUE}) that indicates
weather daraw or not the legend.}
}
\value{
None is returned, only the plot is done.
}
\description{
This function does a radial plot based on the vector of
    letters resulted from pairwise comparisons.
}
\examples{

set.seed(4321)
td <- data.frame(trt = rep(sample(1:20), each = 5))
td$y <- rnorm(nrow(td), mean = 0.15 * sort(td$trt), sd = 1)

plot(y ~ trt, data = td)

# Fit the model.
td$trt <- factor(td$trt)
m0 <- lm(y ~ trt, data = td)
anova(m0)
summary(m0)

library(multcomp)
library(doBy)

X <- LE_matrix(m0, effect = "trt")
rownames(X) <- levels(td$trt)

ci <- apmc(X, m0, focus = "trt", test = "fdr")
ci$cld <- with(ci, ordered_cld(cld, fit))
ci <- ci[order(ci$fit, decreasing = TRUE), ]

library(latticeExtra)

segplot(reorder(trt, fit) ~ lwr + upr,
        centers = fit,
        data = ci,
        draw = FALSE,
        cld = ci$cld) +
    layer(panel.text(x = centers,
                     y = z,
                     labels = sprintf("\%0.2f \%s",
                                      centers,
                                      cld),
                     pos = 3))

radial_cld(cld = ci$cld)
radial_cld(cld = ci$cld, means = ci$fit, perim = TRUE)
radial_cld(cld = ci$cld, col = 1:3)
radial_cld(cld = ci$cld, col = 1:3)
radial_cld(cld = ci$cld, labels = sprintf("\%0.2f \%s", ci$fit, ci$cld))

}
\seealso{
\code{\link{cld2}()}.
}
\author{
Walmes Zeviani, \email{walmes@ufpr.r}.
}
