% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.nls.R
\name{rp.nls}
\alias{rp.nls}
\title{Interactive panel to estimate non linear regression models}
\usage{
rp.nls(model, data, start, subset = NULL, assignTo = "rp.fit",
  startCurve = list(col = 2, lty = 2), fittedCurve = list(col = 2, lty = 1,
  lwd = 1.5), extraplot = NULL, finalplot = NULL, xlab = NULL,
  ylab = NULL, ...)
}
\arguments{
\item{model}{a non linear model formula including variables and
parameters to passed to \code{nls} function.}

\item{data}{a data frame with dependent and independent variables
present in the model formula.}

\item{start}{a named list with one item for each parameter. Itens
names must be equal the parameters. Each item must be named vector
with values for the \code{init}, \code{from} and \code{to}. This is
the order assumed if vector are no named. These values will be used
by \code{rp.slider} as initial, minimum and maximum.}

\item{subset}{an optional string indicating a grouping variable. It
must be a factor.}

\item{assignTo}{the name of the object that will storage the value
returned by the \code{nls} call. It will be a list with one element
per level of the factor used as \code{subset}.}

\item{startCurve}{a list with graphical definitions passed to the
\code{lines} function to draw the curve at the start values. The most
used arguments are \code{lty}, \code{col} and \code{lwd}.}

\item{fittedCurve}{the same as \code{startCurve} but for the
estimated curve.}

\item{extraplot}{a function of the model parameters that draw
auxiliary graphical elements as vertical and horizontal lines as
references for model intercept and/or model asymptote.}

\item{finalplot}{the same as \code{extraplot} but uses the estimated
parameter values.}

\item{xlab}{labels for the x axis.}

\item{ylab}{labels for the y axis.}

\item{...}{arguments passed to the plot function used to draw the
observed values.}
}
\value{
In fact none is returned by the function. There isn't a
\code{return}, instead there is a \code{invisible} at the end. On the
other hand, the model fit is assigned to the object with name passed
to the \code{assingTo} argument.
}
\description{
This function opens a interface, builted with
\code{rpanel} elements, designed for the estimation of non linear
regression models. Sliders allows the user control values used as
start values in the \code{nls} call, checkbox allows select strata of
the data and more options are available.
}
\examples{
\donttest{
##--------------------------------------------
## A simple linear regression.

rp.nls(model=dist~b0+b1*speed,
       data=cars,
       start=list(
           b0=c(init=0, from=-20, to=20),
           b1=c(init=2, from=0, to=10)),
       assignTo="cars.fit")

cars.fit

##--------------------------------------------
## A more interesting example.

require(lattice)
xyplot(rate~conc, groups=state, data=Puromycin)

rp.nls(model=rate~Int+(Top-Int)*conc/(Half+conc),
       data=Puromycin,
       start=list(
           Int=c(init=50, from=20, to=70),
           Top=c(init=150, from=100, to=200),
           Half=c(init=0.1, from=0, to=0.6)),
       subset="state",
       assignTo="Puro.fit",
       startCurve=list(col=3, lty=3, lwd=1),
       fittedCurve=list(col=4, lty=1, lwd=1.5),
       extraplot=function(Int, Top, Half){
           abline(h=c(Int, Top), v=Half, col=2, lty=2)
       },
       finalplot=function(Int, Top, Half){
           abline(h=c(Int, Top), v=Half, col=3, lty=1)
       },
       xlab="Concentration",
       ylab="Rate",
       xlim=c(0, 1.2),
       ylim=c(40, 220))

length(Puro.fit)
sapply(Puro.fit, coef)
sapply(Puro.fit, logLik)
sapply(Puro.fit, deviance)

}
}
\seealso{
\code{\link[stats]{nls}}, \code{\link[graphics]{lines}},
\code{\link[rpanel]{rp.slider}}
}

