% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rp.wire.R
\name{rp.wire}
\alias{rp.wire}
\title{A panel to rotate and choose the better view angle of a
wireframe}
\usage{
rp.wire(wire)
}
\arguments{
\item{wire}{a wireframe object.}
}
\value{
Opens a window with sliders for \code{x}, \code{y} and
\code{z}. Print the \code{dput} os the last view angle using the
button.
}
\description{
A panel rotate a wireframe plot and choose the best view
angle.
}
\examples{
\donttest{
##-----------------------------------------------------------------------------
## A simple example.

require(rpanel)
require(lattice)
require(latticeExtra)
require(RColorBrewer)

colr <- brewer.pal(11, "Spectral")
colr <- colorRampPalette(colr, space="rgb")

grid <- expand.grid(x=seq(-1, 1, by=0.1),
                    y=seq(-1, 1, by=0.1))
grid$z <- with(grid, 1+0.01*x+0.05*y-0.5*x*y-0.5*x^2-0.2*y^2)

p1 <- wireframe(z~x+y, data=grid,
                scales=list(arrows=FALSE),
                ## zlim=extendrange(grid$z, f=0.5),
                ## panel.3d.wireframe=panel.3d.contour,
                ## nlevels=18, col="gray30",
                ## type=c("on", "top", "bottom"),
                col.regions=colr(101),  drape=TRUE)
p1


## Choose the better angle.
rp.wire(p1)

}
}
\seealso{
\code{\link[lattice]{wireframe}}, \code{\link[wzRfun]{panel.3d.contour}}.
}

