% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{soybeanwp}
\alias{soybeanwp}
\title{Factorial experiment to evaluate the effect of potassium and
water on the soybean yield}
\format{A data frame with data from an experiment that evaluated the
effect of water and potassium levels on soybean yield.}
\usage{
data(soybeanwp)
}
\description{
The tropical soils, usually poor in potassium (K),
demand potassium fertilization when cultivated with soybean
(\emph{Glycine max} L.)  to obtain satisfactory yields. The aim of
this study was to evaluate the effects of K doses and soil humidity
levels on soybean agronomic characteristics. The experiment was
carried out in a greenhouse, in pots with two plants, containing 5
dm$^3$ of soil. The experimental design was completely randomized
block with treatments in a 5 x 3 factorial arrangement. The K doses
(\code{potassium}) were 0; 30; 60, 120 and 180 mg dm$^{-3}$ , and the
soil humidity (\code{water}) ranged from 35 to 40; 47.5 to 52.5; and
60 to 65\% of the total porosity (represented as the midpoint). The
characteristics appraised were: grain yield (\code{yield}, grams per
plot), weight of a hundred grains (\code{w100}, grams), total number
of grains per plot (\code{tg}), K concetration in the grain
(\code{Kgrain}, g kg$^{-1}$), number of viable and inviable pods
(\code{nvp, nip}). Observation number 74 can be considered atipical
and removed to do statistical analisys. although the experiment had
been installed in blocks, the greenhouse had constant conditions for
all plots and the management, like irrigation, during the
experimental cicle was carried out with much care, so, the block
effect can be considered null.
}
\details{
This experiment was done at Universidade Federal da Grande
Dourados in Dourados, Mato Grosso do Sul, Brazil. The article based
on this experiment can be acessed by the link
\url{http://ccarevista.ufc.br/seer/index.php/ccarevista/article/view/1454}. Thanks
for the authors by their colaboration and permition to use this data.
}
\keyword{datasets}

