% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_caption}
\alias{table_caption}
\title{Numbered Caption Labels For Exported Tables}
\usage{
table_caption(caption, reset = NULL, label = "Tabela", sep = ":")
}
\arguments{
\item{caption}{a string that is the caption for a table.}

\item{reset}{integer value used to reset the table caption
counter. Default is \code{NULL}.}

\item{label}{a string representing the name for table. Default is
\code{Tabela}.}

\item{sep}{a string with the separator. Default is \code{:}.}
}
\value{
a string composed by \code{label}, \code{sep} and
    \code{caption}. Each call increments its number.
}
\description{
This function does numered caption for exported tables.
}
\examples{

table_caption("First table label.")
table_caption("Second table label.")
table_caption("Third table label.")
table_caption("Counter reseted.", reset = 0)

}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}
\seealso{
\code{\link{table_format}}, \code{\link{matrix2html}}.
}

