% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_format}
\alias{table_format}
\title{Round Numeric Columns of a Table}
\usage{
table_format(table, digits)
}
\arguments{
\item{table}{a data frame.}

\item{digits}{a vector with non negative integer values with length
equals the number of columns of \code{table}. These number
represent the number of digits used to round numeric values. The
number is ignored when the correponding column isn't
numeric. \code{NA} and negative values are ignored.}
}
\value{
a table with column rounded to the number of decimal digits
    specified. All columns are coverted to character, so numeric
    functions can not be direct applied to them anymore. Because of
    this, it is not recommended assign the result of the function to
    the object used as \code{table} argument.
}
\description{
This function returns a table with numeric columns
    rounded accordingly the specified number of decimal
    digits. Numeric columns are coverted to character after rounding.
}
\examples{

x <- table_format(head(rock), digits = c(1, 2, 3, 4))
x
str(x)

x <- table_format(head(iris), c(2, 3, 2, 3, NA))
x
str(x)

x <- table_format(head(rock), c(-1, NA, 3, 4))
x
str(x)

}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}
\seealso{
\code{\link{table_caption}}, \code{\link{matrix2html}}.
}

