% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tableCap}
\alias{tableCap}
\title{Incremental numbered caption labels}
\usage{
tableCap(caption, resetTo = NULL, captionName = "Tabela",
  captionSep = ":")
}
\arguments{
\item{caption}{a string that is the caption for a table.}

\item{resetTo}{integer value used to reset the table caption
counter. Default is \code{NULL}.}

\item{captionName}{a string representing the name for table. Default
is \code{Tabela}.}

\item{captionSep}{a string with the separator. Default is \code{:}.}
}
\value{
a string composed by \code{captionName}, \code{captionSep}
and \code{caption}. Each call increments its number.
}
\description{
This funciton was made to export data frames to html
tables with knitr havind numered captions.
}
\examples{
\donttest{
tableCap("First table label.")
tableCap("Second table label.")
tableCap("Third table label.")
tableCap("Counter reseted.", resetTo=0)
}
}
\seealso{
\code{\link{tableFormat}}, \code{\link{matrix2html}}.
}

