% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tableFormat.R
\name{tableFormat}
\alias{tableFormat}
\title{Round numeric columns of a table with desired number of digits}
\usage{
tableFormat(table, digits)
}
\arguments{
\item{table}{a data frame.}

\item{digits}{a vector with non negative integer values with lenght
equal the number of columns of \code{table}. These number represent
the number of digits used to round numeric values. The number is
ignored when the correponding columns aren't numeric.}
}
\value{
a table with column rounded to the number of decimal digits
specified. All columns are coverted to character, so numeric
functions can not be direct applied to them anymore. Because of this,
it is not recommended assign the result of the function to the object
used \code{table} argument.
}
\description{
This function returns a table with numeric columns
rounded accordingly the specified number of decimal digits. Numeric
columns are coverted to character after rounding.
}
\examples{
\donttest{
x <- tableFormat(head(rock), c(1,2,3,4)); x
str(x)

x <- tableFormat(head(iris), c(2,3,2,3,0)); x
str(x)

tableFormat(head(rock), c(-1,NA,3,4))
tableFormat(as.matrix(head(rock)), c(1,2,3,4))
}
}
\seealso{
\code{\link{tableCap}}, \code{\link{matrix2html}}.
}

