% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree}
\alias{tree}
\title{Show a Directory as an ASCII Tree}
\usage{
tree(path = "./", type = "dir_last", all.files = FALSE)
}
\arguments{
\item{path}{a string indicating the path where the tree starts.}

\item{type}{a string that is type of display. Default is
\code{"dir_last"} to show the directories at the end of the
tree. The alphanumeric order is returned using \code{"none"}.}

\item{all.files}{a logical, if \code{TRUE} hidden files and
directories will be printed.}
}
\value{
Prints on the console the tree of the directory.
}
\description{
This function returns an ascii art of the directory
    tree, like the command
    \href{http://mama.indstate.edu/users/ice/tree/}{\code{tree}} in
    Linux.
}
\examples{

\donttest{

# Tree of your currend directory.
getwd()
tree()
tree(type = "none")
tree(all.files = TRUE)

# Tree of home folder.
tree("~/")

# Tree of a installed package.
tree(system.file(package = "lattice"))

}
}
\seealso{
\code{\link[base]{dir}()}, \code{\link[base]{list.files}()}.
}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}
