% Generated by roxygen2 (4.0.1): do not edit by hand
\name{apmc}
\alias{apmc}
\title{All pairwise mean comparisons}
\usage{
apmc(X, model, focus, test = "single-step", level = 0.05)
}
\arguments{
\item{X}{a matrix where each line is a linear function of the model
parameters to estimate a least squares mean. In most pratical cases,
it is an object from the \code{doBy::LSmatrix()}.}

\item{model}{a model recognized by glht().}

\item{focus}{name of the factor which levels are being compared.}

\item{test}{a p-value correction method.}

\item{level}{the experimentwise level of significance for the
multiple comparisons. The individual coverage of the confidence
interval is \code{1-level}.}
}
\value{
a \code{data.frame} with interval estimates and compact
letter display for the means comparisons.
}
\description{
This function performs all pairwise compararisons among
means returning pontual and intervalar estimates followed by letters
to easy discriminate values.
}
\examples{
\donttest{
require(doBy)
require(multcomp)
## source("apc.R")

## Single factor.
m0 <- lm(weight~feed, data=chickwts)
anova(m0)

X <- LSmatrix(m0, effect="feed")
rownames(X) <- levels(chickwts$feed)
apmc(X, model=m0, focus="feed", test="fdr")

## Two factors.
m1 <- lm(breaks~wool*tension, data=warpbreaks)
anova(m1)

X <- LSmatrix(m1, effect=c("wool", "tension"))
attributes(X)

Xs <- by(X, INDICES=attr(X, "grid")$tension, FUN=as.matrix)
Xs <- lapply(Xs, "rownames<-", levels(warpbreaks$wool))

lapply(Xs, apmc, model=m1, focus="wool", test="single-step",
       level=0.10)
}
}
\seealso{
\code{\link{apc}}, \code{\link[doBy]{LSmatrix}},
\code{\link[multcomp]{glht}}.
}

