% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{apmc}
\alias{apmc}
\title{A Wraper of glht to Get All Pairwise Mean Comparisons}
\usage{
apmc(X, model, focus, test = "single-step", level = 0.05, cld2 = FALSE)
}
\arguments{
\item{X}{a matrix where each line is a linear function of the model
parameters to estimate a least squares mean. In most pratical
cases, it is an object from the \code{\link[doBy]{LSmatrix}()}.}

\item{model}{a model with class recognized by
\code{\link[multcomp]{glht}()}.}

\item{focus}{a string with the name of the factor which levels are
being compared.}

\item{test}{a p-value correction method. See
\code{\link[stats]{p.adjust.methods}()}.}

\item{level}{the experimentwise significance level for the multiple
comparisons. The individual coverage of the confidence interval
is \code{1-level}. Default is 0.05.}

\item{cld2}{Logical, if \code{TRUE} uses the \code{\link{cld2}()}
functions, otherwise uses the \code{\link[multcomp]{cld}()}
function.}
}
\value{
a \code{data.frame} with interval estimates and compact
    letter display for the means comparisons.
}
\description{
This function performs all pairwise compararisons among
    means returning pontual and intervalar estimates followed by
    letters to easy discriminate values. It is in fact a wraper of
    \code{\link[multcomp]{glht}()}.
}
\examples{

library(doBy)
library(multcomp)

# Single factor.
m0 <- lm(weight ~ feed, data = chickwts)
anova(m0)

# Prepare the matrix to estimate lsmeans.
L <- LSmatrix(m0, effect = "feed")
rownames(L) <- levels(chickwts$feed)
apmc(L, model = m0, focus = "feed", test = "fdr")

data(warpbreaks)

# Two factors (complete factorial).
m1 <- lm(breaks ~ wool * tension, data = warpbreaks)
anova(m1)

L <- LSmatrix(m1, effect = c("wool", "tension"))
attributes(L)

Ls <- by(L, INDICES = attr(L, "grid")$tension, FUN = as.matrix)
Ls <- lapply(Ls, "rownames<-", levels(warpbreaks$wool))

# Comparing means of wool in each tension.
lapply(Ls, apmc, model = m1, focus = "wool",
       test = "single-step", level = 0.1)

# Two factors (incomplete factorial).
warpbreaks <- subset(warpbreaks, !(tension == "H" & wool == "A"))
xtabs(~tension + wool, data = warpbreaks)

# There is NA in the estimated parameters.
m2 <- lm(breaks ~ wool * tension, data = warpbreaks)
coef(m2)

X <- model.matrix(m2)
b <- coef(m2)

X <- X[, !is.na(b)]

# unique(X)

# Uses the full estimable model matriz.
m3 <- update(m2, . ~ 0 + X)

# These models are in fact the same.
anova(m2, m3)

# LS matrix has all cells.
L <- LSmatrix(m2, effect = c("wool", "tension"))
g <- attr(L, "grid")
L <- L[, !is.na(b)]
i <- 5
L <- L[-i, ]
g <- g[-i, ]

rownames(L) <- g$tension
Ls <- by(L, INDICES = g$wool, FUN = as.matrix)

# LSmeans with MCP test.
lapply(Ls, apmc, model = m3, focus = "tension",
       test = "single-step", level = 0.1, cld2 = TRUE)

# Sample means.
aggregate(breaks ~ tension + wool, data = warpbreaks, FUN = mean)

}
\seealso{
\code{\link{apc}()}, \code{\link[doBy]{LSmatrix}()},
    \code{\link[multcomp]{glht}()}.
}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}
