% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equallevels.R
\name{equallevels}
\alias{equallevels}
\title{Make Two Data Frames Have The Same Factor Levels}
\usage{
equallevels(x, y)
}
\arguments{
\item{x}{the target data frame.}

\item{y}{the reference data frame.}
}
\value{
the restructured data frame.
}
\description{
If two data frames have factors ou characters columns
    with the same column names, those in the first will have the same
    level order as those in the second. So, for factors, these data
    frames will have the same levels in the same order. Characters in
    the first will be converted to factor.
}
\examples{

a <- data.frame(Species = as.character(
                    sample(iris$Species, size = 20, repl = TRUE)),
                stringsAsFactors = FALSE)
str(a)

b <- equallevels(a, iris)
str(b)

}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}

