% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrix2html.R
\name{matrix2html}
\alias{matrix2html}
\title{Convert a character matrix to HTML table}
\usage{
matrix2html(x, caption = NULL, styDef = NULL, styIndex = NULL,
  tableClass = NULL)
}
\arguments{
\item{x}{a character matrix. Is most of the cases it is the object
returned by \code{tableFormat()}.}

\item{caption}{a string to be caption to the table. Default is
\code{NULL}.}

\item{styDef}{a vector with style definition for \code{th}, \code{tr}
and \code{td} tags. See the examples section. Default is
\code{NULL}. These styles must be defined in a css file used to
render the html page. If you are using \code{knitr}, you can use a
custom css file when knitting
\code{knit2html("my_doc.Rmd", stylesheet="my_custom_css.css")}.}

\item{styIndex}{a positive integer matrix with dimensions equal to
\code{x}. The numbers in each cell of \code{styIndex} call the
corresponding \code{styDef} to be used in \code{x}. See examples for
clarification. Default is \code{NULL}.}

\item{tableClass}{a string corresponding a table style defined in a
css file. Default is \code{NULL}.}
}
\value{
a character vector. Use \code{cat()} inside chunks with
\code{results="asis"} to print the result to be interpreted by a
browser as an html table.
}
\description{
This function covert a character matrix to html table.
}
\examples{
\donttest{
x <- head(rock)
x <- sapply(x, as.character)
str(x)
m2h <- matrix2html(x)
cat(m2h)

x <- mtcars[1:10, 1:3]
x <- rbind(c("", colnames(x)), cbind(rownames(x), x))
x <- as.matrix(x)
rownames(x) <- colnames(x) <- NULL
x[1,1] <- "Cars"
x

m2h <- matrix2html(x, caption="Part of the cars data set.")
cat(m2h)

styIndex <- 0L*row(x)+3L
styIndex[3,] <- 4L
styIndex[4,] <- 5L
styIndex[5,] <- 6L
styIndex[7,] <- 7L
styIndex[8,] <- 8L
styIndex[9,] <- 9L
styIndex[,3] <- 3L
styIndex[1,] <- 1L

## These table class are assumed to be defined in a css file as
## follow.

## td.red { color: #CC0000; }
## td.blue { color: #0000CC; }
## td.shade { background-color: #CCCCCC; }
## td.line { border-bottom: 1px solid #000000; }
## td.bold { font-weight: bold; }
## td.italic { font-style: italic; }

styDef <- c("<th align=\\"center\\">\%s</th>", "<tr>\\n\%s\\n</tr>\\n",
            "<td align=\\"center\\">\%s</td>",
            "<td align=\\"center\\" class=\\"red\\">\%s</td>",
            "<td align=\\"center\\" class=\\"blue line\\">\%s</td>",
            "<td align=\\"center\\" class=\\"shade\\">\%s</td>",
            "<td align=\\"center\\" class=\\"line\\">\%s</td>",
            "<td align=\\"center\\" class=\\"bold shade\\">\%s</td>",
            "<td align=\\"center\\" class=\\"italic blue\\">\%s</td>")

m2h <- matrix2html(x=x, styDef=styDef, styIndex=styIndex,
                   caption="Part of the cars data set.")
cat(m2h)
}
}
\seealso{
\code{\link{tableFormat}}, \code{\link{tableCap}}.
}

