% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepanel.cbH.R
\name{prepanel.cbH}
\alias{prepanel.cbH}
\title{Lattice panels for error bars and envelop bands}
\usage{
prepanel.cbH(y, ly, uy, subscripts)
}
\arguments{
\item{y}{central value (point estimate).}

\item{ly}{lower limit.}

\item{uy}{upper limit.}

\item{subscripts}{see \code{\link[lattice]{panel.xyplot}}}
}
\value{
None is returned.
}
\description{
Used to do or add error bars or envelop bands in
plots. It is used in with \code{\link{panel.cbH}}.
}
\seealso{
\code{\link{panel.cbH}}
}

