% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree}
\alias{tree}
\title{Show directories as indented names}
\usage{
tree(path = ".", options = "-F --charset=ascii", linux = TRUE)
}
\arguments{
\item{path}{character[1] Is the root directory. Default is the value
of \code{getwd()}.}

\item{options}{character[1] String of options to be passed to
\code{tree} shell command. Default is
\code{"-F --charset=ascii"}. See the available options with
\code{system("tree --help")}.}

\item{linux}{logical[1] Default is \code{TRUE} and indicates that
\code{tree} of the Linux SO is to be used. Otherwise, the R
native approach will be used.}
}
\value{
Prints an ASCII art of the directory tree.
}
\description{
Show directories as indented names
}
\details{
This function calls
    \href{http://mama.indstate.edu/users/ice/tree/}{\code{tree}} if
    it is available is Linux OS. The call is made with
    \code{system()}. Otherwise, a (less sophisticated) R native
    approach is used.
}
\examples{

\dontrun{

tree()
tree("../")
tree("../", linux = FALSE)

system("tree --help")

tree(system.file(package = "lattice"),
     options = "-F --charset=ascii -L 1")
tree(system.file(package = "lattice"),
     options = "-F --charset=ascii -L 2")

}

}
\seealso{
\code{link{dir_tree}} for an implementation entirely based
    on R and functional in all SO. Also, \code{\link[base]{dir}},
    \code{\link[base]{list.files}} and \code{\link[base]{list.dirs}}
    can be useful in many circumstances.
}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}
